pyproject.toml
[project]
name = "bedrock-kb-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.36.9",
    "langchain-aws>=0.2.11",
    "langchain-community>=0.3.16",
    "python-dotenv>=1.0.1",
]
